Move drawing code into own mod
